Split legacy DOM text overlays
